Move JSON response classes to JSON subdirectory
