+++ src/Http/Response/JSON/BadRequest.php
@@ -1,7 +1,7 @@
 <?php declare(strict_types=1);
 /**
- * File contains the PSR15 ResponseInterface implementation for
- * Not Modified responses.
+ * File contains the PSR7 ResponseInterface implementation for
+ * Bad Request responses.
  *
  * PHP Version 7
  *
@@ -14,13 +14,13 @@
  * @see https://www.php-fig.org/psr/psr-7/
  * @see https://www.php-fig.org/psr/psr-15/
  */
-namespace LORIS\Http\Response;
+namespace LORIS\Http\Response\JSON;
 
 use \LORIS\Http\Response\JsonResponse;
 
 /**
  * A LORIS Http Response is an implementation of the PSR15 ResponseInterface
- * to use in LORIS specific for 304 Not Modified.
+ * to use in LORIS specific for 400 Bad Request.
  *
  * @category PSR15
  * @package  Http
@@ -28,16 +28,16 @@ use \LORIS\Http\Response\JsonResponse;
  * @license  http://www.gnu.org/licenses/gpl-3.0.txt GPLv3
  * @link     https://www.github.com/aces/Loris/
  */
-class NotModified extends JsonResponse
+class BadRequest extends JsonResponse
 {
     /**
-     * Create a Json response specific to 404 Not Found
+     * Create a Json response specific to 400 Bad Request
      *
-     * @param string $etag The endpoint etag
+     * @param string $msg The error message
      */
-    public function __construct(string $etag)
+    public function __construct(string $msg = 'bad request')
     {
-        $headers = array('ETag' => $etag);
-        parent::__construct(null, 304, $headers);
+        $body = array('error' => $msg);
+        parent::__construct($body, 400);
     }
 }
