+++ src/Http/Response/JSON/MethodNotAllowed.php
@@ -1,7 +1,7 @@
 <?php declare(strict_types=1);
 /**
  * File contains the PSR15 ResponseInterface implementation for
- * Not Implemented responses.
+ * Method Not Allowed responses.
  *
  * PHP Version 7
  *
@@ -14,13 +14,13 @@
  * @see https://www.php-fig.org/psr/psr-7/
  * @see https://www.php-fig.org/psr/psr-15/
  */
-namespace LORIS\Http\Response;
+namespace LORIS\Http\Response\JSON;
 
 use \LORIS\Http\Response\JsonResponse;
 
 /**
  * A LORIS Http Response is an implementation of the PSR15 ResponseInterface
- * to use in LORIS specific for 501 Not Implemented.
+ * to use in LORIS specific for 405 Method Not Allowed.
  *
  * @category PSR15
  * @package  Http
@@ -28,16 +28,23 @@ use \LORIS\Http\Response\JsonResponse;
  * @license  http://www.gnu.org/licenses/gpl-3.0.txt GPLv3
  * @link     https://www.github.com/aces/Loris/
  */
-class NotImplemented extends JsonResponse
+class MethodNotAllowed extends JsonResponse
 {
     /**
-     * Create a Json response specific to 501 Not Implemented
+     * The server MUST generate an Allow header field in a 405 response
+     * containing a list of the target resource's currently supported methods.
      *
-     * @param string $msg The error message
+     * @param array  $allowedmethods The list of supported methods
+     * @param string $msg            The error message
+     *
+     * @see RFC 7231, section 6.5.5: 405 Method Not Allowed
      */
-    public function __construct(string $msg = 'not implemented')
-    {
-        $body = array('error' => $msg);
-        parent::__construct($body, 501);
+    public function __construct(
+        array $allowedmethods,
+        string $msg = 'method not allowed'
+    ) {
+        $body    = array('error' => $msg);
+        $headers = array('Allow' => $allowedmethods);
+        parent::__construct($body, 405, $headers);
     }
 }
