+++ src/Http/Response/JSON/NotFound.php
@@ -1,7 +1,7 @@
 <?php declare(strict_types=1);
 /**
  * File contains the PSR15 ResponseInterface implementation for
- * Internal Server Error responses.
+ * Not Found responses.
  *
  * PHP Version 7
  *
@@ -14,13 +14,13 @@
  * @see https://www.php-fig.org/psr/psr-7/
  * @see https://www.php-fig.org/psr/psr-15/
  */
-namespace LORIS\Http\Response;
+namespace LORIS\Http\Response\JSON;
 
 use \LORIS\Http\Response\JsonResponse;
 
 /**
  * A LORIS Http Response is an implementation of the PSR15 ResponseInterface
- * to use in LORIS specific for 500 Internal Server Error.
+ * to use in LORIS specific for 404 Not Found.
  *
  * @category PSR15
  * @package  Http
@@ -28,16 +28,16 @@ use \LORIS\Http\Response\JsonResponse;
  * @license  http://www.gnu.org/licenses/gpl-3.0.txt GPLv3
  * @link     https://www.github.com/aces/Loris/
  */
-class InternalServerError extends JsonResponse
+class NotFound extends JsonResponse
 {
     /**
-     * Create a Json response specific to 500 Internal Server Error
+     * Create a Json response specific to 404 Not Found
      *
      * @param string $msg The error message
      */
-    public function __construct(string $msg = 'internal server error')
+    public function __construct(string $msg = 'not found')
     {
         $body = array('error' => $msg);
-        parent::__construct($body, 500);
+        parent::__construct($body, 404);
     }
 }
