+++ src/Http/Response/JSON/Unauthorized.php
@@ -1,7 +1,7 @@
 <?php declare(strict_types=1);
 /**
  * File contains the PSR15 ResponseInterface implementation for
- * Method Not Allowed responses.
+ * Unauthorized responses.
  *
  * PHP Version 7
  *
@@ -14,13 +14,13 @@
  * @see https://www.php-fig.org/psr/psr-7/
  * @see https://www.php-fig.org/psr/psr-15/
  */
-namespace LORIS\Http\Response;
+namespace LORIS\Http\Response\JSON;
 
 use \LORIS\Http\Response\JsonResponse;
 
 /**
  * A LORIS Http Response is an implementation of the PSR15 ResponseInterface
- * to use in LORIS specific for 405 Method Not Allowed.
+ * to use in LORIS specific for 401 Unauthorized.
  *
  * @category PSR15
  * @package  Http
@@ -28,23 +28,16 @@ use \LORIS\Http\Response\JsonResponse;
  * @license  http://www.gnu.org/licenses/gpl-3.0.txt GPLv3
  * @link     https://www.github.com/aces/Loris/
  */
-class MethodNotAllowed extends JsonResponse
+class Unauthorized extends JsonResponse
 {
     /**
-     * The server MUST generate an Allow header field in a 405 response
-     * containing a list of the target resource's currently supported methods.
+     * Create a Json response specific to 401 Unauthorized
      *
-     * @param array  $allowedmethods The list of supported methods
-     * @param string $msg            The error message
-     *
-     * @see RFC 7231, section 6.5.5: 405 Method Not Allowed
+     * @param string $msg The error message
      */
-    public function __construct(
-        array $allowedmethods,
-        string $msg = 'method not allowed'
-    ) {
-        $body    = array('error' => $msg);
-        $headers = array('Allow' => $allowedmethods);
-        parent::__construct($body, 405, $headers);
+    public function __construct(string $msg = 'unauthorized')
+    {
+        $body = array('error' => $msg);
+        parent::__construct($body, 401);
     }
 }
